+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.0] - 2024-01-24
+
+### Added
+
+- `-p` flag to display only lines that were correctly parsed. No other log formats shall pass.
+- Tests in the github workflows
+
 ## [0.10.1] - 2024-01-23
 
 ### Added